Minor fixes to README.rst
There are two changes here. The first is a clarification about what should be
passed to Betfair when setting up the client login by changing test to app_key.
My second change is a correction to the number of events returned for a search
on 'tennis'. The search returns two results as table tennis also matches the
filter in the example below.