Implement betfair.price.ticks_difference(..)
This methods accepts two exact, valid Betfair prices and returns the
absolute difference in number of "ticks" between them, e.g.:

    ticks_between(1.01, 1.01) == 0
    ticks_between(1.01, 1.02) == 1
    ticks_between(1.03, 1.01) == 2

i.e. the result is always a non-negative integer regardless of which
price is bigger than the other.

If one or both of the give prices is not an exact, valid Betfair price
then this method will throw an exception.