Implement betfair.price.price_ticks_away(..)
This methods accepts an exact, valid Betfair prices with a number of
"ticks" and returns the price that number of ticks away, e.g:

    price_ticks_away(1.01, 0) == 1.01
    price_ticks_away(1.01, 1) == 1.02
    price_ticks_away(1.02, -1) == 1.01

If the given price is not an exact, valid Betfair price then this
method will throw an exception. An exception will also be thrown if
there is no price the given number of ticks away.