Avoid repeated field name serialization
Previously the betfair.model.ModelMeta metaclass added an attribute
_fields to BetfairModel, this was a dict mapping field name to type for
each Field declared on model subclasses. As the field names are in
snake case (in Python) but camel case (in Betfair API JSON responses)
the serialize and unserialize methods of BetfairModel had to convert
between the two.

This changes ModelMeta to add the _fields attribute with "serialized"
(i.e. the same as in API responses) names as the keys of the dict. This
way the names only have to be converted from snake case to camel case
once.