Remove betfair.model.BetfairModel.unserialize_key
As the _fields attribute of BetfairModel instances now has "serialized"
field names as keys there is no need to "unserialize" keys in API JSON
responses to determine which attribute to set on the model. This method
was just been used to log a warning message when a JSON object contained
a key for which there was no field in the model. Now the name of the JSON
key is logged instead.