Remove "license" argument from setup
"license" is only meant to be used if there isn't a classifier
available for the license the package is under, there is already one
for the MIT license and setup.py already declares it.

Additionaly, the LICENSE file is not packaged with the rest of the
library, and so `read('LICENSE')` fails when attempting to install
from pypi (meaning you cannot install 0.1.4 using
`pip install betfair.py` currently).